Remove unecessary state. Simplify code.
